+++ lib/generators/css_zero/install/install_generator.rb
@@ -14,7 +14,7 @@ class CssZero::InstallGenerator < Rails::Generators::Base
   end
 
   def add_basic_components
-    invoke "css_zero:add", %w(layouts flash alert input button)
+    invoke "css_zero:add", %w(layouts flash alert input button icons)
   end
 
   private